Rename and move decorators
