+++ connexion/decorators/lifecycle.py
